Refactoring and introducing internal namespaces
