+++ OrbitCaptureClient/CaptureEventProcessorProcessEventsFuzzer.cpp
@@ -24,8 +24,7 @@ class MyCaptureListener : public CaptureListener {
   void OnCaptureStarted(
       int32_t /*process_id*/, std::string /*process_name*/, std::shared_ptr<Process> /*process*/,
       absl::flat_hash_map<uint64_t, orbit_client_protos::FunctionInfo> /*selected_functions*/,
-      absl::flat_hash_set<orbit_grpc_protos::TracepointInfo, HashTracepointInfo,
-                          EqualTracepointInfo> /*selected_tracepoints*/) override {}
+      TracepointInfoSet /*selected_tracepoints*/) override {}
   void OnCaptureComplete() override {}
   void OnTimer(const TimerInfo&) override {}
   void OnKeyAndString(uint64_t, std::string) override {}
