+++ OrbitCaptureClient/include/OrbitCaptureClient/CaptureListener.h
@@ -5,11 +5,11 @@
 #ifndef ORBIT_CAPTURE_CLIENT_CAPTURE_LISTENER_H_
 #define ORBIT_CAPTURE_CLIENT_CAPTURE_LISTENER_H_
 
-#include "../../OrbitGl/TracepointCustom.h"
 #include "Callstack.h"
 #include "EventBuffer.h"
 #include "OrbitProcess.h"
 #include "ScopeTimer.h"
+#include "TracepointCustom.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "capture_data.pb.h"
@@ -22,9 +22,7 @@ class CaptureListener {
   virtual void OnCaptureStarted(
       int32_t process_id, std::string process_name, std::shared_ptr<Process> process,
       absl::flat_hash_map<uint64_t, orbit_client_protos::FunctionInfo> selected_functions,
-      absl::flat_hash_set<orbit_grpc_protos::TracepointInfo, HashTracepointInfo,
-                          EqualTracepointInfo>
-          selected_tracepoints) = 0;
+      TracepointInfoSet selected_tracepoints) = 0;
   // Called when capture is complete
   virtual void OnCaptureComplete() = 0;
