+++ OrbitClientGgp/ClientGgp.cpp
@@ -81,8 +81,7 @@ bool ClientGgp::RequestStartCapture(ThreadPool* thread_pool) {
 
   // Start capture
   LOG("Capture pid %d", pid);
-  absl::flat_hash_set<orbit_grpc_protos::TracepointInfo, HashTracepointInfo, EqualTracepointInfo>
-      selected_tracepoints;
+  TracepointInfoSet selected_tracepoints;
 
   ErrorMessageOr<void> result =
       capture_client_->StartCapture(thread_pool, pid, target_process_->GetName(), target_process_,
@@ -229,8 +228,7 @@ absl::flat_hash_map<uint64_t, FunctionInfo> ClientGgp::GetSelectedFunctions() {
 void ClientGgp::OnCaptureStarted(
     int32_t process_id, std::string process_name, std::shared_ptr<Process> process,
     absl::flat_hash_map<uint64_t, orbit_client_protos::FunctionInfo> selected_functions,
-    absl::flat_hash_set<orbit_grpc_protos::TracepointInfo, HashTracepointInfo, EqualTracepointInfo>
-        selected_tracepoints) {
+    TracepointInfoSet selected_tracepoints) {
   capture_data_ =
       CaptureData(process_id, process_name, process, selected_functions, selected_tracepoints);
   LOG("Capture started");
