+++ OrbitGl/CMakeLists.txt
@@ -55,8 +55,7 @@ target_sources(
          TimerTrack.h
          Track.h
          TriangleToggle.h
-         TracepointsDataView.h
-         TracepointCustom.h)
+         TracepointsDataView.h)
 
 target_sources(
   OrbitGl
