+++ OrbitGl/DataManager.cpp
@@ -154,7 +154,4 @@ bool DataManager::IsTracepointSelected(const TracepointInfo& info) const {
   return selected_tracepoints_.contains(info);
 }
 
-const absl::flat_hash_set<TracepointInfo, HashTracepointInfo, EqualTracepointInfo>&
-DataManager::selected_tracepoints() const {
-  return selected_tracepoints_;
-}
+const TracepointInfoSet& DataManager::selected_tracepoints() const { return selected_tracepoints_; }
